// src/api/models/ClientCreateSuccessDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientCreateSuccessDto = {
    success: boolean;
    message: string;
    method: Record<string, any>;
    endpoint: string;
    statusCode: number;
    timestamp: string;
    data?: Record<string, any>;
};