// src/api/models/ClientDeleteSuccessDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientDeleteSuccessDto = {
    success: boolean;
    message: string;
    method: Record<string, any>;
    endpoint: string;
    statusCode: number;
    timestamp: string;
    data?: Record<string, any>;
};